requirements.txt
streamlit>=1.24.0
pandas>=1.5.0,<2.2.0
plotly>=5.13.0
requests>=2.28.0
python-dotenv>=0.21.0
openpyxl>=3.0.10
sendgrid>=6.9.0
numpy>=1.22.0